llm_project/requirements.txt
django==5.1.3
djangorestframework=3.15.2
llama-cpp-python=0.3.1
langchain-core=0.3.15
langchain-community=0.3.5
langchain-huggingface=0.1.2
langchain-text-splitters=0.3.2
langchain-weaviate=0.0.3
pdfplumber=0.11.4
six=1.16.0
weaviate-client=4.9.3